feat(ClassTable): 教室預覽、向度/學年、代碼複製，並修正 0.5 學分被算成 0

+++ swift/TigerDuck/Features/ClassTable/ClassTableViewModel.swift
@@ -322,7 +322,7 @@ final class ClassTableViewModel {
         TigerDuckTheme.ensureAssignments(courseNos: Array(courseNos))
     }
 
-    var totalCredits: Int {
+    var totalCredits: Double {
         courses.reduce(0) { $0 + $1.credits }
     }
